feat(admin): add integrated academics dashboard
